рефакторинг вывода статистики
